Change usage to receive total alloc / hole size
Current usage, which is receiving number of allocations and number of
holes, is easy to use but requires additional calculation to user to
adjust total size of allocation and hole.  This commit eliminates the
additional calculation from user by changing usage.  Now, we receive
total size of allocation and total size of entire holes.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>